Implement 'FldDungeon_CreateTask'

+++ include/Model/mdlManager.h
@@ -127,9 +127,9 @@ typedef struct Model
 
 extern const f32 gFrameDuration;
 
-Model* MdlManager_InitMdl(u32 type, u32 id);
-Model* MdlManager_CreateMdl(u32 type, u32 id, const char* path, u32 flags);
-Model* MdlManager_CreateMdl2(u32 type, u32 id, u32 flags);
+Model* MdlManager_InitMdl(u16 type, u16 id);
+Model* MdlManager_CreateMdl(u16 type, u16 id, const char* path, u32 flags);
+Model* MdlManager_CreateMdl2(u16 type, u16 id, u32 flags);
 
 f32 MdlAnim_GetDurationInFrame(Model* mdl, u16 slotIdx);
 f32 MdlAnim_GetDurationInFrameById(Model* mdl, u16 slotIdx, s16 animId);
